backend/requirements.txt
fastapi==0.111.0
uvicorn==0.30.1
beanie==1.25.0
motor==3.4.0
pymongo==4.6.3
python-dotenv==1.0.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
fastapi-mail==1.4.1
python-multipart==0.0.9
email-validator==2.1.1
jinja2==3.1.4
pydantic-settings==2.2.1
